requirements.txt
Django
django-sass-processor
Pillow
channels
daphne
django-debug-toolbar
beautifulsoup4
requests
django-cors-headers
mysqlclient
djangorestframework